refactor(orchestrator): extract SharingMemberRecordResolutionService and toUuidOrNull utility to eliminate duplication
