+++ bpdm-orchestrator/src/main/kotlin/org/eclipse/tractusx/bpdm/orchestrator/service/parser/GoldenRecordTaskCreateParser.kt
@@ -21,25 +21,27 @@ package org.eclipse.tractusx.bpdm.orchestrator.service.parser
 
 import org.eclipse.tractusx.bpdm.common.model.ParseResult
 import org.eclipse.tractusx.bpdm.common.model.zipParseResults
-import org.eclipse.tractusx.bpdm.orchestrator.entity.SharingMemberRecordDb
 import org.eclipse.tractusx.bpdm.orchestrator.model.error.GoldenRecordTaskCreateParseError
 import org.eclipse.tractusx.bpdm.orchestrator.model.parsed.GoldenRecordTaskCreateParsed
 import org.eclipse.tractusx.bpdm.orchestrator.model.request.BusinessPartnerRequest
 import org.eclipse.tractusx.bpdm.orchestrator.model.request.GoldenRecordTaskCreateRequest
-import org.eclipse.tractusx.bpdm.orchestrator.repository.SharingMemberRecordRepository
+import org.eclipse.tractusx.bpdm.orchestrator.service.SharingMemberRecordResolutionService
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import java.util.UUID
 
 @Service
 @Transactional(readOnly = true)
 class GoldenRecordTaskCreateParser(
-    private val sharingMemberRecordRepository: SharingMemberRecordRepository
+    private val sharingMemberRecordResolutionService: SharingMemberRecordResolutionService
 ) {
 
     fun parse(requests: List<GoldenRecordTaskCreateRequest>): List<ParseResult<GoldenRecordTaskCreateParsed, GoldenRecordTaskCreateParseError>> {
         val businessPartnerResults = requests.map { validateBusinessPartner(it.businessPartner) }
-        val gateRecordResults = resolveGateRecords(requests.map { it.recordId })
+        val gateRecordResults = sharingMemberRecordResolutionService.resolveGateRecords(
+            requests.map { it.recordId },
+            { GoldenRecordTaskCreateParseError.RecordIdInvalid(it) },
+            { GoldenRecordTaskCreateParseError.RecordNotFound(it) }
+        )
 
         return zipParseResults(businessPartnerResults, gateRecordResults) { businessPartner, gateRecord ->
             GoldenRecordTaskCreateParsed(existingGateRecord = gateRecord, businessPartner = businessPartner)
@@ -51,28 +53,4 @@ class GoldenRecordTaskCreateParser(
             ParseResult.ofSingleFailure(GoldenRecordTaskCreateParseError.AdditionalSitesWithoutSite)
         else
             ParseResult.Success(businessPartner)
-
-    private fun resolveGateRecords(recordIds: List<String?>): List<ParseResult<SharingMemberRecordDb?, GoldenRecordTaskCreateParseError>> {
-        val parsedUuids = recordIds.map { recordId -> recordId?.let { toUuidOrNull(it) } }
-
-        val recordsByPrivateId = parsedUuids.filterNotNull().toSet()
-            .let { sharingMemberRecordRepository.findByPrivateIdIn(it) }
-            .associateBy { it.privateId }
-
-        return recordIds.zip(parsedUuids).map { (recordId, uuid) ->
-            when {
-                recordId == null -> ParseResult.Success(null)
-                uuid == null -> ParseResult.ofSingleFailure(GoldenRecordTaskCreateParseError.RecordIdInvalid(recordId))
-                else -> recordsByPrivateId[uuid]?.let { ParseResult.Success(it) }
-                    ?: ParseResult.ofSingleFailure(GoldenRecordTaskCreateParseError.RecordNotFound(recordId))
-            }
-        }
-    }
-
-    private fun toUuidOrNull(uuidString: String): UUID? =
-        try {
-            UUID.fromString(uuidString)
-        } catch (_: IllegalArgumentException) {
-            null
-        }
 }
